+++ php/session.php
@@ -9,7 +9,6 @@ namespace bbsengine6\session;
 
 require_once("/srv/www/bbsengine6/php/bootstrap.php");
 
-/// require_once("config.php");
 require_once("engine.php");
 require_once("database.php");
 require_once("libmember.php");